requirements.txt
django==4.2.15
django-quill-editor
channels[daphne]
pycrdt-websocket